@@ -7,31 +7,34 @@
     "start": "node ./node_modules/@angular/cli/bin/ng serve --proxy-config proxy.conf.json",
     "build": "node ./node_modules/@angular/cli/bin/ng build --prod",
     "test": "node ./node_modules/@angular/cli/bin/ng test",
-    "test-single-run": "node ./node_modules/@angular/cli/bin/ng test --single-run",
+    "test-single-run": "node ./node_modules/@angular/cli/bin/ng test --watch=false",
     "lint": "node ./node_modules/@angular/cli/bin/ng lint",
     "e2e": "node ./node_modules/@angular/cli/bin/ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.x",
-    "@angular/common": "5.2.x",
-    "@angular/compiler": "5.2.x",
-    "@angular/core": "5.2.x",
-    "@angular/forms": "5.2.x",
-    "@angular/http": "5.2.x",
-    "@angular/platform-browser": "5.2.x",
-    "@angular/platform-browser-dynamic": "5.2.x",
-    "@angular/router": "5.2.x",
+    "@angular/animations": "7.2.x",
+    "@angular/common": "7.2.x",
+    "@angular/compiler": "7.2.x",
+    "@angular/core": "7.2.x",
+    "@angular/forms": "7.2.x",
+    "@angular/http": "7.2.x",
+    "@angular/platform-browser": "7.2.x",
+    "@angular/platform-browser-dynamic": "7.2.x",
+    "@angular/router": "7.2.x",
     "bootstrap": "4.1.x",
     "core-js": "2.5.x",
     "intl": "1.2.x",
-    "rxjs": "5.5.x",
-    "zone.js": "0.8.x"
+    "rxjs": "6.4.x",
+    "tslib": "1.9.x",
+    "zone.js": "0.8.x",
+    "rxjs-compat": "^6.0.0-rc.0"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.3",
-    "@angular/compiler-cli": "5.2.x",
-    "@angular/language-service": "5.2.x",
+    "@angular-devkit/build-angular": "0.13.x",
+    "@angular/cli": "7.3.x",
+    "@angular/compiler-cli": "7.2.x",
+    "@angular/language-service": "7.2.x",
     "@types/jasmine": "2.8.x",
     "@types/jasminewd2": "2.0.x",
     "@types/node": "9.6.x",
@@ -49,6 +52,6 @@
     "ts-mockito": "2.3.x",
     "ts-node": "5.0.x",
     "tslint": "5.9.x",
-    "typescript": "2.7.x"
+    "typescript": "3.2.x"
   }
 }
